Add validation against a list of emails
Allows an email or list of emails to be used for restricting access

Use cases:
- Smaller setups where you just need to overlay oauth for a
  handful of random gmails.
- Situations where you need to enable singular email addresses outside
  your domain

Whats changed:
Adds new "email" flag similar to the existing domain flag. Csv enabled.

This is then validated in a similar way to the domain but in an
additive fashion allowing domain validation to be checked, but if that
fails and the email matches it will still pass.

Alternative Option:
I considered replacing domain validation or adding a new option for a
regex but felt this met the use cases I have quicker and more cleanly.