Add a google group validation
In the same way that the email domain and the email whitelist is checked, the
email is now checked for google group membership. To check an email is a member
of a google group requires access to the gcloud API for the whole google domain.
The official Google Gcloud API client library is used and requires extra
configuration:

- GoogleDomain is the domain managed by GCloud
- GoogleGroup is a comma delimited list of the required group emails
- GoogleApplicationCredentials is path to service account JSON credentials file
- GoogleActingAdminEmail is the email an admin account with domain access
- GoogleExpirySeconds is the number of seconds the API response is cached

The service account requires domain-wide delegation of authority as discussed
here https://developers.google.com/admin-sdk/directory/v1/guides/delegation.
It also requires the oauth scope
https://www.googleapis.com/auth/admin.directory.group.readonly to be able to
read the groups of a member email. This service account has to impersonate an
admin account to access the admin APIs.

The default google groups required for authorization is specified in the
GoogleGroup variable, but can also be specified in a Rule like:

--rule.1.action=allow
--rule.1.rule=Host(`example.com`)
--rule.1.google_groups=team@example.com

The order of the rules is significant and controls the priority of the rule
matching.