Add support for testing OIDC groups via the rules engine for RBAC
This adds support for rules like

  rules.1.groups=sysadmins,superusers
  rules.1.groupmode=any|all

This looks for a `groups` claim on the token, which is not a part of the
spec but provided by multiple providers.

The serialized cookie is now a b64-encoded JSON payload, which attaches
group information to the user. Validation includes comparing the user's
authorized groups to the rules configured on the server.

A groupmode of "any" (default) will permit the request if any of the given groups is
matched. A mode of "all" will permit the request only if ALL the given groups are matched.